Fix click on login, go, search options
Fix the click events on login, go, and search options buttons.

By using preventDefault, we avoid the click triggering a reload of
the page/app.

Various onclick= attributes were removed on the
buttons, and the click handlers are set up by the document ready
function in main.js.

Since the Login button is created in a number of places, we use a
dedicated init_login_onclick function to set the click handler.

Signed-off-by: Jeff Godin <jgodin@tadl.org>