Include mobi_bypass argument in JSON urls
Work around some situations where (due to user agent) we were being
redirected to an HTML page instead of receiving the JSON data that
we requested.

This manifested itself in Featured / Locations / Events not working
the first time one was selected, but then all further attempts to
load the data (during a single run of the app) were successful.

The mobi_bypass argument ensures that our current method of mobile
redirection does not fire on the server side.

Signed-off-by: Jeff Godin <jgodin@tadl.org>