Fix URL for locations data
Fix URL for locations data -- I appended a query string to
LOCATIONS_BASE, which later (as the _BASE implies) has additional
data appended to it.

This resulted in an incorrect URL, and the Locations button would
never display data.

Rename variable to LOCATIONS_URL and include full URL with no later
appending.

NOTE: We've worked around this on the server side with a
RewriteRule, so this does not require an immediate new release of
the app.

Signed-off-by: Jeff Godin <jgodin@tadl.org>