Mitigate input issues when scrolled down
Mitigate issues where login elements would disappear from view when
the user attempts to log in after performing a search and scrolling
down in the list of results.

Most of this seems to be caused by UIWebView issues with
position:fixed when the soft keyboard is active.

iOS 5.1.1 quirk: the first tap in the input field will bring up the
keyboard but not give focus to the input field.

Signed-off-by: Jeff Godin <jgodin@tadl.org>